refactored Console class
